+++ Vault/src/utopia/vault/database/columnlength/ColumnLengthLimits.scala
@@ -2,6 +2,7 @@ package utopia.vault.database.columnlength
 
 import utopia.flow.async.Volatile
 import utopia.flow.datastructure.immutable.DeepMap
+import utopia.vault.model.immutable.{Column, Table}
 
 /**
   * Used for tracking length limits applying to each database column
@@ -13,7 +14,7 @@ object ColumnLengthLimits
 	// ATTRIBUTES   ------------------------------
 	
 	// (database name -> table name -> property name) => column length limit
-	private val limits2 = Volatile(DeepMap.empty[String, ColumnLengthLimit])
+	private val limits = Volatile(DeepMap.empty[String, ColumnLengthLimit])
 	
 	
 	// OTHER    ----------------------------------
@@ -25,18 +26,45 @@ object ColumnLengthLimits
 	  * @return A length limit to apply to that column, if specified
 	  */
 	def apply(databaseName: String, tableName: String, propertyName: String) =
-		limits2.value.get(databaseName, tableName, propertyName)
+		limits.value.get(databaseName, tableName, propertyName)
+	/**
+	  * @param databaseName Name of targeted database
+	  * @param column Targeted column
+	  * @return A length limit to apply to that column, if specified
+	  */
+	def apply(databaseName: String, column: Column): Option[ColumnLengthLimit] =
+		apply(databaseName, column.tableName, column.name)
+	/**
+	  * @param table Targeted table
+	  * @param propertyName Name of the targeted property within that table
+	  * @return A length limit to apply to that column / property, if applicable
+	  */
+	def apply(table: Table, propertyName: String): Option[ColumnLengthLimit] =
+		apply(table.databaseName, table.name, propertyName)
+	/**
+	  * @param databaseName Name of targeted database
+	  * @param tableName Name of targeted table
+	  * @return A map that contains length limits for property names where applicable
+	  */
+	def apply(databaseName: String, tableName: String) =
+		limits.value.nested(databaseName, tableName).flat
+	/**
+	  * @param table Targeted table
+	  * @return Column length limits to apply to that table
+	  */
+	def apply(table: Table): Map[String, ColumnLengthLimit] = apply(table.databaseName, table.name)
+	
 	/**
 	  * @param key A key consisting of database name, table name and property name
 	  * @param limit A limit to apply to that column
 	  */
 	def update(key: (String, String, String), limit: ColumnLengthLimit) =
-		limits2.update { _ + (Vector(key._1, key._2, key._3) -> limit) }
+		limits.update { _ + (Vector(key._1, key._2, key._3) -> limit) }
 	/**
 	  * @param databaseName Name of targeted database
 	  * @param tableName Name of targeted table
 	  * @param lengthLimits Length limits to apply, where keys are property names and values are length limits
 	  */
 	def update(databaseName: String, tableName: String, lengthLimits: IterableOnce[(String, ColumnLengthLimit)]) =
-		limits2.update { _ ++ (Vector(databaseName, tableName) -> DeepMap.flat(lengthLimits)) }
+		limits.update { _ ++ (Vector(databaseName, tableName) -> DeepMap.flat(lengthLimits)) }
 }